Trait btleplug::api::bleuuid::BleUuid

pub trait BleUuid {
    fn to_ble_u32(&self) -> Option<u32>;
    fn to_ble_u16(&self) -> Option<u16>;
    fn to_short_string(&self) -> String;
}

An extension trait for Uuid which provides BLE-specific methods.

Required methods

fn to_ble_u32(&self) -> Option<u32>

If the UUID is a valid BLE short UUID then return its short form, otherwise return None.

fn to_ble_u16(&self) -> Option<u16>

If the UUID is a valid 16-bit BLE short UUID then return its short form, otherwise return None.

fn to_short_string(&self) -> String

Convert the UUID to a string, using short format if applicable.